@@ -1,15 +1,17 @@
 {
-  "name": "marxan_web",
-  "version": "v1.0.6",
+  "name": "bioprotect_web",
+  "version": "v1.10.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-regular-svg-icons": "^5.12.0",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
-    "@fortawesome/react-fontawesome": "^0.1.8",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
     "@mapbox/mapbox-gl-draw": "^1.1.1",
-    "@material-ui/core": "^4.0.0",
-    "@material-ui/icons": "^4.11.3",
+    "@mui/base": "^5.0.0-beta.40",
+    "@mui/icons-material": "^5.15.21",
+    "@mui/material": "^5.15.21",
     "axios": "^1.7.2",
     "classybrew": "0.0.3",
     "fetch-jsonp": "^1.1.3",
@@ -18,14 +20,11 @@
     "mapbox": "^1.0.0-beta10",
     "mapbox-gl": "^1.11.0",
     "rc-progress": "^2.5.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-fontawesome": "^1.7.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hexagon": "^1.1.3",
-    "react-scripts": "^5.0.1",
     "react-stay-scrolled": "^3.0.0",
-    "react-table": "^6.11.5",
-    "react-textarea-autosize": "^7.1.2",
+    "react-table": "^7.8.0",
     "recharts": "^2.12.7",
     "typeface-roboto": "0.0.54",
     "wellknown": "^0.5.0"
@@ -40,7 +39,10 @@
   },
   "homepage": "https://app.marxanweb.org",
   "devDependencies": {
-    "esformatter-jsx": "^7.2.0"
+    "@babel/plugin-transform-private-property-in-object": "^7.24.7",
+    "esformatter-jsx": "^7.2.0",
+    "react-scripts": "^5.0.1",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11" 
   },
   "browserslist": {
     "production": [
