@@ -1,11 +1,11 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 // import { faSync } from '@fortawesome/free-solid-svg-icons';
-import Import from "@material-ui/icons/GetApp";
+import Import from "@mui/icons-material/GetApp";
 import MarxanDialog from "./MarxanDialog";
 import MarxanTable from "./MarxanTable";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import Popover from "@material-ui/core/Popover";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import Popover from "@mui/material/Popover";
 /*
  * Copyright (c) 2020 Andrew Cottam.
  *
@@ -17,12 +17,12 @@ import Popover from "@material-ui/core/Popover";
 import React from "react";
 import TableRow from "./TableRow.js";
 import ToolbarButton from "./ToolbarButton";
-import { faCheckCircle } from "@fortawesome/free-regular-svg-icons";
-import { faCircle } from "@fortawesome/free-regular-svg-icons";
+import { faCheckCircle } from "@fortawesome/free-solid-svg-icons";
+import { faCircle } from "@fortawesome/free-solid-svg-icons";
 import { faPlusCircle } from "@fortawesome/free-solid-svg-icons";
-import { faTrashAlt } from "@fortawesome/free-regular-svg-icons";
+import { faTrashAlt } from "@fortawesome/free-solid-svg-icons";
 //import gbif_logo from './gbif.jpg';
-//import FontIcon from '@material-ui/core/FontIcon';
+//import FontIcon from '@mui/material/FontIcon';
 // add this to the gbif item: leftIcon={<FontIcon className="gbifLogo"/>
 
 class FeaturesDialog extends React.Component {
