@@ -1,7 +1,7 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import MarxanDialog from "./MarxanDialog";
 import MarxanTextField from "./MarxanTextField";
-import MenuItem from "@material-ui/core/MenuItem";
+import MenuItem from "@mui/material/MenuItem";
 /*
  * Copyright (c) 2020 Andrew Cottam.
  *
@@ -11,7 +11,7 @@ import MenuItem from "@material-ui/core/MenuItem";
  * License: European Union Public Licence V. 1.2, see https://opensource.org/licenses/EUPL-1.2
  */
 import React from "react";
-import Select from "@material-ui/core/Select";
+import Select from "@mui/material/Select";
 import { faLock } from "@fortawesome/free-solid-svg-icons";
 import { faUnlink } from "@fortawesome/free-solid-svg-icons";
 
