@@ -9,10 +9,10 @@
 import * as React from "react";
 
 import CONSTANTS from "./constants";
-import Checkbox from "@material-ui/core/Checkbox";
+import Checkbox from "@mui/material/Checkbox";
 import MarxanDialog from "./MarxanDialog";
-import MenuItem from "@material-ui/core/MenuItem";
-import Select from "@material-ui/core/Select";
+import MenuItem from "@mui/material/MenuItem";
+import Select from "@mui/material/Select";
 import ToolbarButton from "./ToolbarButton";
 
 class NewProjectWizardDialog extends React.Component {
