@@ -1,3 +1,5 @@
+import MarxanDialog from "./MarxanDialog";
+import MarxanTable from "./MarxanTable";
 /*
  * Copyright (c) 2020 Andrew Cottam.
  *
@@ -7,9 +9,6 @@
  * License: European Union Public Licence V. 1.2, see https://opensource.org/licenses/EUPL-1.2
  */
 import React from "react";
-import "react-table/react-table.css";
-import MarxanDialog from "./MarxanDialog";
-import MarxanTable from "./MarxanTable";
 
 class AtlasLayersDialog extends React.Component {
   constructor(props) {
