@@ -1,3 +1,6 @@
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import MarxanDialog from "../MarxanDialog";
+import MarxanTable from "../MarxanTable";
 /*
  * Copyright (c) 2020 Andrew Cottam.
  *
@@ -7,13 +10,9 @@
  * License: European Union Public Licence V. 1.2, see https://opensource.org/licenses/EUPL-1.2
  */
 import React from "react";
-import "react-table/react-table.css";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faPlusCircle } from "@fortawesome/free-solid-svg-icons";
-import ToolbarButton from "../ToolbarButton";
-import MarxanDialog from "../MarxanDialog";
-import MarxanTable from "../MarxanTable";
 import TableRow from "../TableRow.js";
+import ToolbarButton from "../ToolbarButton";
+import { faPlusCircle } from "@fortawesome/free-solid-svg-icons";
 
 class ImportedActivitiesDialog extends React.Component {
   constructor(props) {
@@ -60,9 +59,10 @@ class ImportedActivitiesDialog extends React.Component {
     if (this.state.selectedUploadedActivityIds.includes(rowInfo.original.id)) {
       // remove the rowInfo.original
       this.setState((prevState) => ({
-        selectedUploadedActivityIds: prevState.selectedUploadedActivityIds.filter(
-          (act) => act !== rowInfo.original.id
-        ),
+        selectedUploadedActivityIds:
+          prevState.selectedUploadedActivityIds.filter(
+            (act) => act !== rowInfo.original.id
+          ),
       }));
     } else {
       // add the rowInfo.original id
