@@ -1,3 +1,5 @@
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import MarxanDialog from "./MarxanDialog";
 /*
  * Copyright (c) 2020 Andrew Cottam.
  *
@@ -7,10 +9,7 @@
  * License: European Union Public Licence V. 1.2, see https://opensource.org/licenses/EUPL-1.2
  */
 import React from "react";
-import "react-table/react-table.css";
 import ReactTable from "react-table";
-import FontAwesome from "react-fontawesome";
-import MarxanDialog from "./MarxanDialog";
 
 class RunSettingsDialog extends React.Component {
   constructor(props) {
